refactor(frontend): split monolithic pages into composable sub-components
- Extract shared UI primitives (StatusBadge, StatCard, EmptyState, ErrorBanner,
  SuccessBanner, SectionHeader) into components/ui/ for cross-page reuse
- Split governance/page.tsx (670→130 lines): types, constants, helpers,
  useGovernanceState hook, and 8 focused sub-components (PolicyMetaPanel,
  FujiRulesEditor, DiffPreview, RiskImpactGauge, ApprovalWorkflow, ApplyFlow,
  TrustLogStream, ChangeHistory)
- Split audit/page.tsx (716→130 lines): AuditSummaryPanel, AuditTimeline,
  VerificationPanel, ExportPanel extracted with clear prop interfaces
- Split risk/page.tsx (705→130 lines): types, constants, data-helpers,
  useRiskStream hook, and 6 focused sub-components (RiskScatterPlot,
  InsightCards, TrendChart, FlaggedRequestsList, DrilldownPanel, WhyFlaggedPanel)
- All 153 tests pass, zero new lint warnings, no type errors introduced
- Preserves all existing behavior, API contracts, i18n, and a11y
